refactoring and imrovements
* refactored log2 stuff - everything now in log2.h/log2.cpp
* minor fixed hinted by clang-tidy
* added largemul (backport) --> still buggy (test fails)
* still bugs in bitreverse --> test fails
* removed unused files

+++ the_framework/unittest/utest_dbllnkdlst.cpp
@@ -129,7 +129,7 @@ TEST_CASE("Datastructures::dbllnkdlst"){
       });
       CAPTURE(ii);
       REQUIRE( !found.first );
-      REQUIRE( (ii-1) == found.second );
+      REQUIRE( u16(ii-1) == found.second );
     }
   }
 
@@ -148,7 +148,7 @@ TEST_CASE("Datastructures::dbllnkdlst"){
       });
       CAPTURE(ii);
       REQUIRE( !found.first );
-      REQUIRE( (ii-1) == found.second );
+      REQUIRE( u16(ii-1) == found.second );
     }
   }
